Use env variable for API token
