+++ .gitignore
@@ -1,5 +1,6 @@
 __pycache__/
 hawaii-et-rain-website/__pycache__/
+hawaii-et-rain-website/*.db
 plots/
 models/
 helpers/
