+++ main.py
@@ -7,7 +7,9 @@ from math import sqrt
 from datetime import datetime
 import concurrent.futures
 
-API_TOKEN = "1b8a6439c85b8e42e211b68ea68ac198"  # Replace with your actual HCDP API token
+API_TOKEN = os.getenv("HCDP_API_TOKEN")
+if not API_TOKEN:
+    raise RuntimeError("HCDP_API_TOKEN environment variable not set")
 farms = [
     {"name": "Kahuku Farm", "lat": 21.6832, "lng": -157.9604},
     {"name": "Nozawa Farms", "lat": 21.688, "lng": -157.9648},
